endboss finished

+++ classes/world.class.js
@@ -43,7 +43,7 @@ class World {
     checkCollisions() {
             this.level.enemies.forEach((enemy) => {
                 if(this.character.isColliding(enemy)) {
-                    if(!this.character.isDead() && !this.character.isHurt()) {
+                    if(!this.character.isDead() && !this.character.isHurt() && !enemy.isDead()) {
                         this.character.hurt(enemy.damage);
                     }
                 }
@@ -67,7 +67,7 @@ class World {
                     endboss.setHealthbar();
                 }
                 this.throwables.forEach((bottle => {
-                    if(endboss.isColliding(bottle)) {
+                    if(endboss.isColliding(bottle) && !endboss.isDead()) {
                         endboss.hurt(20);      
                         bottle.delete();                  
                     }
